backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
django-ratelimit==4.1.0
bleach==6.1.0
django-environ==0.11.2